fix: Refactor codigo de relations y setup por conflicto de metodos
